Params class for generating the url query params
This class will take in all of the options and allow us to get a hash of
valid options. These will then be passed to Addressable to generate the
URL. All error handling/formatting will be handled here.

Params with nil values are fine because Addressable will just ignore
them. :)