Always send times in UTC
The AddThisEvent requires the timezone to be in location format (aka
'Africa/Casablanca').

It's impossible to get this value accurately from a Ruby DateTime object
(might be off because of daylight savings time).

To make this work. We'll sent all times as UTC and tell AddThisEvent the
timezone is 'Africa/Casablanca'. Which is 00:00 neutral zone with no
DST. (I confirmed with AddThisEvent that this would correctly map to
UTC).